requirements.txt
numpy
matplotlib 
Pillow
opencv-python
torch
torchvision



